requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.104.1
h11==0.14.0
idna==3.4
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.1
pandas==2.1.2
passlib==1.7.4
psycopg2-binary==2.9.9
py-bcrypt==0.4
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
